requirements.txt
Flask
openai
python-dotenv
# Hinweis build prozess fuer lxml dauert ewig, funktioniert aber